add : 로그인이 필요한 회원수정, 회원탈퇴 기능에 기능을 사용하는 유저가 맞는지 확인하는 로직 추가[#6]

+++ src/main/java/me/moon/ticketReservation/customer/service/CustomerService.java
@@ -1,11 +1,14 @@
 package me.moon.ticketReservation.customer.service;
 
 import lombok.RequiredArgsConstructor;
+import me.moon.ticketReservation.common.exception.ErrorCode;
 import me.moon.ticketReservation.customer.dto.CustomerResponseDto;
 import me.moon.ticketReservation.customer.dto.CustomerSaveRequestDto;
 import me.moon.ticketReservation.customer.dto.CustomerUpdateRequestDto;
 import me.moon.ticketReservation.customer.entity.Customer;
 import me.moon.ticketReservation.customer.repository.CustomerMapper;
+import me.moon.ticketReservation.login.entity.SessionUser;
+import me.moon.ticketReservation.login.exception.UnauthorizedAccessException;
 import me.moon.ticketReservation.supplier.exception.DuplicateEmailException;
 import me.moon.ticketReservation.supplier.exception.WithdrawalException;
 import org.springframework.stereotype.Service;
@@ -30,14 +33,25 @@ public class CustomerService {
         return customerMapper.isDuplicateEmail(email);
     }
 
-    public CustomerResponseDto update(String customerId, CustomerUpdateRequestDto dto) {
+    public CustomerResponseDto update(SessionUser sessionUser, String customerId, CustomerUpdateRequestDto dto) {
         Customer customer = customerMapper.findById(customerId);
+        if ( !compareUserAndSessionUser(sessionUser, customer) ) {
+            throw new UnauthorizedAccessException(sessionUser.getEmail(), ErrorCode.UNAUTHORIZED_ACCESS);
+        }
         customer.update(dto);
         customerMapper.update(customer);
         return CustomerResponseDto.of(customer);
     }
 
-    public void delete(String customerId) {
+    private boolean compareUserAndSessionUser(SessionUser sessionUser, Customer customer) {
+        return sessionUser.getEmail().equals(customer.getEmail());
+    }
+
+    public void delete(SessionUser sessionUser, String customerId) {
+        Customer customer = customerMapper.findById(customerId);
+        if ( !compareUserAndSessionUser(sessionUser, customer) ) {
+            throw new UnauthorizedAccessException(sessionUser.getEmail(), ErrorCode.UNAUTHORIZED_ACCESS);
+        }
         int result = customerMapper.deleteById(customerId);
         if ( result != 1 ) {
             throw new WithdrawalException(customerId);
